cache/requirements.txt
streamlit==1.20.0
python-dotenv==1.0.0
openai==0.27.0
whisper==1.0
pytube==15.0.0
ffmpeg-python==0.2.0
streamlit-authenticator

